% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atwt2d.R
\name{atwt2d}
\alias{atwt2d}
\title{Computes A trous wavelet transform (ATWT)}
\usage{
atwt2d(data2d, max_scale = -1)
}
\arguments{
\item{data2d}{2d image as array or matrix.}

\item{max_scale}{computes wavelets up to \code{max_scale}. Leave blank for maximum possible scales.}
}
\value{
array containing ATWT of input image with added 3rd dimention for scales.
}
\description{
Computes ATWT of the 2d array up to \code{max_scale}.
If \code{max_scale} is outside the boundaries, number of scales will be reduced.
Data is mirrored at the boundaries.'Negative WT are removed. Not tested for non-square data.
}
\note{
Need to break this into smaller functions.
}
\author{
Dileep M. Puranik and Bhupendra Raut
}
