% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt-class.R
\name{class3dTo2d}
\alias{class3dTo2d}
\title{2D projection of 3D convective-stratiform classes.}
\usage{
class3dTo2d(wt_class_3d, vert_clust, vert_range = 1:30)
}
\arguments{
\item{wt_class_3d}{Volume classification obtained from \code{\link{get_class}}}

\item{vert_clust}{Clusters saved in RDs files from functions \code{\link{clusterProfiles}}}

\item{vert_range}{same vert_range in as \code{\link{saveVertProf}} and \code{\link{sampleVertProf}}}
}
\value{
2d array of pixels labeled with three classes. 1. stratiform, 2. Convection, 3. Mixed (Need correction)
}
\description{
Checks verticle profile of the classification and finds continuous
regions of similar classification and assigns one dominent class.
If both classes has comparable presence, mixed class is assigned.
}
\seealso{
\code{\link{get_class}}
}
