% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt-class.R
\name{cleanWT}
\alias{cleanWT}
\title{Remove tiny fluctuations that may not be of interest.}
\usage{
cleanWT(wt_sum, dbz_vol)
}
\arguments{
\item{wt_scan}{2d wt image at a scale or sum of several scales.}

\item{times_sd}{Default value=1. pixels with WT value < mean + times_sd * SD are removed.}
}
\value{
wt_scan with small values removed.
}
\description{
This may not be needed for clean dataset.
Use when WT has too much noise with trial and error approach.
}
