% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust-prof.R
\name{clusterProfiles}
\alias{clusterProfiles}
\title{saves clusters of verticle profiles to disk.}
\usage{
clusterProfiles(vprof_sample, nclust = 5)
}
\arguments{
\item{vprof_sample}{saved samples from the function \code{\link{saveVertProf}}}

\item{nclust}{Desired number of clusters. (Default 5)}
}
\description{
This function clusters given verticle profiles using K-Modes algorithm 
and saves it in RDS file.
}
\seealso{
\url{https://cran.r-project.org/web/packages/klaR/index.html}{klaR} package

\code{\link{saveVertProf}}
}
