% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-io.R
\name{create_outNC_class}
\alias{create_outNC_class}
\title{creates output netCDF file usin variables and dims from Steiner classification file.}
\usage{
create_outNC_class(outfPath, time_seconds, sample_file)
}
\description{
This is done to keep consistency between the files for easier comparison.
I dont know if this is still valid.
}
