% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt-class.R
\name{dbz2rr}
\alias{dbz2rr}
\title{computes rain rate using Z-R relationship.}
\usage{
dbz2rr(dbz, ZRA = 200, ZRB = 1.6)
}
\arguments{
\item{dbz}{array, vector or matrix of reflectivity in dBZ}
}
\value{
rr rain rate in \code{mm/hr}
}
\description{
Uses standard values for ZRA=200 and ZRB=1.6.
}
