% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-io.R
\name{get_1dayFiles}
\alias{get_1dayFiles}
\title{gives list of files belong to the same day as the first file in list_allFiles.}
\usage{
get_1dayFiles(list_allFiles, datestr_revpos)
}
\arguments{
\item{list_allFiles}{output of Glob or list.files()}

\item{datestr_revpos}{the position of date in file name from the reverse direction.}
}
\description{
Method: split the first file's name by "_" and extract 'dates' string,
then search for the same 'dates' in the list and return them back.
}
