% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-io.R
\name{get_nctime}
\alias{get_nctime}
\title{reads time from a single netcdf file. Using with laply.}
\usage{
get_nctime(filename)
}
\arguments{
\item{filename}{file name}
}
\value{
time in POSIX standard with "1970-01-01" as origin.
}
\description{
Titan/CF radial generated files do not have constatnt time epoch. The time units changes for every file.
Therefore, I had to write this complicated function  to convert variable time epoch to standard unix epoch.
}
\note{
This function has sevarl assumptions about the way time units are written. 
It may not work with files other than CF-radial.
}
