% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atwt2d.R
\name{getMaxScale}
\alias{getMaxScale}
\title{Calculate the mximum possible scale of ATWT for given dimensions.}
\usage{
getMaxScale(data_dims)
}
\arguments{
\item{data_dim}{output of the \code{dim(data2d)} for given matrix or array.}
}
\value{
integer value of the maximum scale.
}
\description{
Calculate the mximum possible scale of ATWT for given dimensions.
}
