% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt-class.R
\name{getWTClass}
\alias{getWTClass}
\title{Compute scan-by-scan ATWT of radar volume.}
\usage{
getWTClass(dbz_data, res_km, conv_scale = 20)
}
\arguments{
\item{\code{vol_data}}{3D array containing radar data. Last dimension should be levels.}

\item{\code{res_km}}{resolution of the radar data in km.}

\item{\code{conv_scale}}{approximate scale break (in km) between convective and stratiform scales.}
}
\value{
Sum of wavelets upto \code{conv_scale} for each scan.
}
\description{
Converts dBZ to rain rates using standard Z-R relationship.
This is to transform the normally distributed dBZ to gamma-like distribution.
}
\seealso{
\code{\link{getWTSum}}
}
