% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt-class.R
\name{labelClasses}
\alias{labelClasses}
\title{Lables 1. stratiform, 2. intense/heavy convective  and 3. moderate+transitional convective regions 
using given thersholds.}
\usage{
labelClasses(wt_sum, vol_data)
}
\description{
Lables 1. stratiform, 2. intense/heavy convective  and 3. moderate+transitional convective regions 
using given thersholds.
}
