% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust-prof.R
\name{sampleVertProf}
\alias{sampleVertProf}
\title{sample verticle arrangement of Co and non-Co levels in the txt file for clustering.}
\usage{
sampleVertProf(wt_class_3d, n = 10, vert_range = 1:30)
}
\arguments{
\item{wt_class_3d}{Volume of WT classification from \code{\link{getWTClass}}}

\item{n}{only few profiles will be saved from this volume for efficiency. default=10.}
}
\description{
This function is required to get the verticle structure modes using k-modes algorithm.
The function is not used in actual classification run.
}
\seealso{
\code{link{nearest_kmode}} function.
}
