% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust-prof.R
\name{saveVertProf}
\alias{saveVertProf}
\title{Saves verticle profiles to disk}
\usage{
saveVertProf(fout = "./profiles.txt", dbz_vol, scale_break,
  vert_range = 1:30, nsample = 10)
}
\arguments{
\item{fout}{output file path}

\item{dbz_vol}{3d radar reflectvity volume that will be used for classification.}

\item{scale_break}{scale break obtained from \code{\link{getScaleBreak}}}

\item{vert_range}{number of verticle levels to be considered for classification}

\item{nsample}{only few random veticle samples will taken from each image. (Default 10)}
}
\description{
Saves verticle profiles to disk
}
\note{
Provide only rainy volume scans to this function. Keep \code{nsample} small.
}
