% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt-class.R
\name{dbz2rr}
\alias{dbz2rr}
\title{computes rain rate using standard Z-R relationship.}
\usage{
dbz2rr(dbz)
}
\arguments{
\item{dbz}{array, vector or matrix of reflectivity in dBZ}
}
\value{
rr rain rate in \code{mm/hr}
}
\description{
computes rain rate using standard Z-R relationship.
}
